[fix] - correcao exercicio02

+++ exercicio01/src/parte01/Main01.java
@@ -18,7 +18,7 @@ public class Main01 {
         int jan = 15000;
         int fev= 23000;
         int mar= 17000;
-        int despTotal = jan+fev+mar;
+        double despTotal = jan+fev+mar;
         
         System.out.println("Despesa total: "+ despTotal);     
         System.out.println("Média mensal: "+ despTotal/3);
